Исправлено предупреждение в компиляторе IAR AVR.
git-svn-id: https://5-10/svn/mxsrclib/trunk@712 63356120-0428-6b47-bcd6-cd59724df219

+++ irssysutils.h
@@ -254,7 +254,7 @@ void number_to_string(const T& a_num, string* ap_str)
     (type_to_index<T>() == signed_char_idx) ||
     (type_to_index<T>() == unsigned_char_idx))
   {
-    int val_int = a_num;
+    int val_int = static_cast<int>(a_num);
     ostr << val_int << irst('\0');
   } else {
     ostr << a_num << irst('\0');
